% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmc.R
\name{cmc_to_month}
\alias{cmc_to_month}
\title{Convert century-month-code to a month}
\usage{
cmc_to_month(cmc)
}
\arguments{
\item{cmc, }{the number of months since the beginning of the 20th century.
January 1900 is CMC 1.}
}
\description{
Convert century-month-code to a month
}
