% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-home.R
\name{examples}
\alias{examples}
\title{Returns the package's examples directory (for the loaded
version).}
\usage{
examples()
}
\description{
Returns the package's examples directory (for the loaded
version).
}
