% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-metadata.R
\name{extract_column_labels}
\alias{extract_column_labels}
\title{Extract column labels to match column codes.}
\usage{
extract_column_labels(data)
}
\arguments{
\item{data}{data frame with embedded column
labels in 'label' attribute.}
}
\value{
named vector with names being codes and
        values being labels.
}
\description{
Extract column labels to match column codes.
}
