% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-filenames.R
\name{extract_country}
\alias{extract_country}
\title{Extract country name from file name.}
\usage{
extract_country(file)
}
\arguments{
\item{file}{name to extract from.}
}
\value{
return the long-form country name used by DHS.
}
\description{
Extract country name from file name.
}
