% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-codes.R
\name{extract_country_code}
\alias{extract_country_code}
\title{For a filename 'file', extract the country code (first two characters).}
\usage{
extract_country_code(file)
}
\arguments{
\item{file}{filename to extract from.}
}
\description{
For a filename 'file', extract the country code (first two characters).
}
