% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-filenames.R
\name{extract_dataset_type}
\alias{extract_dataset_type}
\title{Extract the dataset type from the file name.}
\usage{
extract_dataset_type(file)
}
\arguments{
\item{file}{name to extract from.}
}
\value{
return the long-form of the datatype used by DHS.
}
\description{
Extract the dataset type from the file name.
}
