% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-codes.R
\name{extract_dataset_type_code}
\alias{extract_dataset_type_code}
\title{For a filename 'file', extract the dataset type code (3rd and 4th characters).}
\usage{
extract_dataset_type_code(file)
}
\arguments{
\item{file}{filename to extract from.}
}
\description{
For a filename 'file', extract the dataset type code (3rd and 4th characters).
}
