% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-filenames.R
\name{extract_dhs_release}
\alias{extract_dhs_release}
\title{Extract the DHS round's data release version from the file name.}
\usage{
extract_dhs_release(file)
}
\arguments{
\item{file}{name to extract from.}
}
\value{
return an alphanumeric from the DHS file name. Usually a digit.
}
\description{
Extract the DHS round's data release version from the file name.
}
