% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-codes.R
\name{extract_dhs_release_code}
\alias{extract_dhs_release_code}
\title{For a filename 'file', extract the within-round data release code (6th character).}
\usage{
extract_dhs_release_code(file)
}
\arguments{
\item{file}{filename to extract from.}
}
\description{
For a filename 'file', extract the within-round data release code (6th character).
}
