% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-filenames.R
\name{extract_dhs_round}
\alias{extract_dhs_round}
\title{Extract the DHS round from the file name.}
\usage{
extract_dhs_round(file)
}
\arguments{
\item{file}{name to extract from.}
}
\value{
return a digit for the DHS round (e.g.-DHS-III -> 3)
}
\description{
Extract the DHS round from the file name.
}
