% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-codes.R
\name{extract_dhs_round_code}
\alias{extract_dhs_round_code}
\title{For a filename 'file', extract the DHS round code (5th character).}
\usage{
extract_dhs_round_code(file)
}
\arguments{
\item{file}{filename to extract from.}
}
\description{
For a filename 'file', extract the DHS round code (5th character).
}
