% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{extract_dta_files}
\alias{extract_dta_files}
\title{Extract .dta files from DHS .zip files and insert data on
extracted .dta files into the appropriate file_index elements.
Files are copied to the output path all lower-case as is
our convention.}
\usage{
extract_dta_files(file_index, output_path)
}
\arguments{
\item{file_index}{holding info on zip files and contents
generated by, at least, `list_files`.}

\item{output_path}{where extracted .dta files are written.}
}
\value{
file_index updated with an 'extracted_file` element
        for each entry that contained a .dta file.
}
\description{
Extract .dta files from DHS .zip files and insert data on
extracted .dta files into the appropriate file_index elements.
Files are copied to the output path all lower-case as is
our convention.
}
