% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-filenames.R
\name{extract_format}
\alias{extract_format}
\title{Extract the file format from the file name.}
\usage{
extract_format(file)
}
\arguments{
\item{file}{name to extract from.}
}
\value{
return the long-form of the file format
}
\description{
Extract the file format from the file name.
}
