% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-codes.R
\name{extract_format_code}
\alias{extract_format_code}
\title{For a filename 'file', extract the data format code (7th and 9th characters).}
\usage{
extract_format_code(file)
}
\arguments{
\item{file}{filename to extract from.}
}
\description{
For a filename 'file', extract the data format code (7th and 9th characters).
}
