% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-codes.R
\name{extract_survey_stub}
\alias{extract_survey_stub}
\title{For a filename 'file', extract the stub (first 8 characters)}
\usage{
extract_survey_stub(file)
}
\arguments{
\item{file}{filename to extract from.}
}
\description{
For a filename 'file', extract the stub (first 8 characters)
}
