% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-metadata.R
\name{extract_value_labels}
\alias{extract_value_labels}
\title{Extract column value labels to match column value codes.}
\usage{
extract_value_labels(data)
}
\arguments{
\item{data}{data frame with embedded column
value labels in 'labels' attribute.}
}
\value{
named vector with names being codes and
        values being labels.
}
\description{
Extract column value labels to match column value codes.
}
