% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-files.R
\name{filter_file_name}
\alias{filter_file_name}
\title{Filter a path for DHS data file name elements.  Returns 
true for path items that match the requested elements and
false for those that don't.  The match considers the 
position of the element in the filename so it's not a straight
regex.}
\usage{
filter_file_name(path, country = TRUE, dataset = TRUE, round = TRUE,
  release = TRUE, format = TRUE)
}
\arguments{
\item{path}{single path or vector of paths to filter against.}

\item{country}{country CODE (not name) to filter with.}

\item{dataset}{dataset type CODE to filter with.}

\item{round}{round number (e.g.-3 for DHS-III).}

\item{release}{release version code (not always numeric, varies 
by round and country). This function should really support
'latest' as a special value.}

\item{format}{file format CODE (not extension or name) to filter
with.}
}
\description{
To avoid situations with unreliable filtering (e.g.-try typing
"Lao People's Democratic Republic" reliably/repeatedly) this
function relies on the DHS codes for countries, data formats, 
etc... to get these use the functions 'get_country_code', 
'get_dataset_type_code', 'get_file_format_extensions', 
'get_file_format_code'.  Those functions to loose string matching
via regex.
}
