% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-files.R
\name{filtered_unzip}
\alias{filtered_unzip}
\title{Unzip from a path to a single file or multiple files to a specified output
output directory but only if the internal files has a given extension.}
\usage{
filtered_unzip(path, extension = NULL, output_dir = getwd())
}
\arguments{
\item{path}{path to a file or vector of files to paths to unzip from.}

\item{extension}{extension to filter for, no filtering if left unspecified.}

\item{output_dir}{directory to write unzipped files to.}
}
\description{
Unzip from a path to a single file or multiple files to a specified output
output directory but only if the internal files has a given extension.
}
