% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-metadata.R
\name{get_column_code}
\alias{get_column_code}
\title{Get column code matching a search string}
\usage{
get_column_code(s, data, ...)
}
\arguments{
\item{s}{string to search for}

\item{data}{data frame with embeded column labels}

\item{...}{additional arguments to regex search.}
}
\description{
Get column code matching a search string
}
