% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-metadata.R
\name{get_column_label}
\alias{get_column_label}
\title{Get column label matching a column code}
\usage{
get_column_label(c, data)
}
\arguments{
\item{c}{code to search for}

\item{data}{data frame with embeded column labels}
}
\description{
Get column label matching a column code
}
