% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-codes.R
\name{get_country_code}
\alias{get_country_code}
\title{Retrieves the country code based on a 
text match against 's'.}
\usage{
get_country_code(s, ...)
}
\arguments{
\item{s}{string to match}

\item{...}{additional arguments passed to grepl,
e.g.-ignore.case=TRUE}
}
\value{
the country code or codes matching the string
}
\description{
Retrieves the country code based on a 
text match against 's'.
}
