% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-maps.R
\name{get_country_code_map}
\alias{get_country_code_map}
\title{Return a hacked-together named vector connecting country
DHS codes to country names.}
\usage{
get_country_code_map()
}
\value{
a vector with codes for names and country names for values.
}
\description{
Return a hacked-together named vector connecting country
DHS codes to country names.
}
