% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-maps.R
\name{get_country_code_page}
\alias{get_country_code_page}
\title{This downloads the page for the URL that _currently_ has the 
DHS country codes... could probably look up what standard they
are using but IDK right now.}
\usage{
get_country_code_page()
}
\value{
an XML of the web page containing this info.
}
\description{
This downloads the page for the URL that _currently_ has the 
DHS country codes... could probably look up what standard they
are using but IDK right now.
}
