% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-codes.R
\name{get_dataset_type_code}
\alias{get_dataset_type_code}
\title{Retrieves the dataset type code based on a 
text match against 's'.}
\usage{
get_dataset_type_code(s, ...)
}
\arguments{
\item{s}{string to match}

\item{...}{additional arguments passed to grepl,
e.g.-ignore.case=TRUE}
}
\value{
the dataset type code or codes matching the string
}
\description{
Retrieves the dataset type code based on a 
text match against 's'.
}
