% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-maps.R
\name{get_dataset_type_map}
\alias{get_dataset_type_map}
\title{Return a hackeed-together named vector connecting country 
DHS dataset types to their descriptions.}
\usage{
get_dataset_type_map()
}
\value{
a vector with codes for names and descriptions for values.
}
\description{
Return a hackeed-together named vector connecting country 
DHS dataset types to their descriptions.
}
