% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-maps.R
\name{get_dataset_type_page}
\alias{get_dataset_type_page}
\title{This downloads the page for the URL that _currently_ has the 
DHS dataset type codes.}
\usage{
get_dataset_type_page()
}
\value{
an XML of the web page containing this info.
}
\description{
This downloads the page for the URL that _currently_ has the 
DHS dataset type codes.
}
