% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-codes.R
\name{get_file_format_code}
\alias{get_file_format_code}
\title{Retrieves DHS file format code based on a text
match against 's'.}
\usage{
get_file_format_code(s, ...)
}
\arguments{
\item{s}{string to match}

\item{...}{additional arguments passed to grepl, 
e.g.-ignore.case=TRUE}
}
\value{
the file format code or codes matching the string
}
\description{
Retrieves DHS file format code based on a text
match against 's'.
}
