% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-codes.R
\name{get_file_format_extensions}
\alias{get_file_format_extensions}
\title{Retrieves the file format extension based on
a text match against 's'.}
\usage{
get_file_format_extensions(s, ...)
}
\arguments{
\item{s}{string to match}

\item{...}{additional arguments passed to grepl,
e.g.-ignore.case=TRUE}
}
\value{
the file extension or extensions matching the string
}
\description{
Retrieves the file format extension based on
a text match against 's'.
}
