% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-maps.R
\name{get_file_format_map}
\alias{get_file_format_map}
\title{Return a named vector describing data formats and
their DHS codes.}
\usage{
get_file_format_map()
}
\value{
named vector with data format codes for names and
        format names for values.
}
\description{
Return a named vector describing data formats and
their DHS codes.
}
