% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-scripts.R
\name{get_function}
\alias{get_function}
\title{Retrieve the implied name of the main function for a script.}
\usage{
get_function(job, logger)
}
\value{
function name as character vector, length-1
}
\description{
Retrieve the implied name of the main function for a script.
}
