% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-recodes.R
\name{get_manual_paths}
\alias{get_manual_paths}
\title{Retrieves the links to the recode manuals for all 
rounds.}
\usage{
get_manual_paths()
}
\value{
url for each recode manual and the associated map .pdf
}
\description{
Retrieves the links to the recode manuals for all 
rounds.
}
