% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-recode-manuals.R
\name{get_recode_manual_page}
\alias{get_recode_manual_page}
\title{Current location of the DHS recode manuals.}
\usage{
get_recode_manual_page()
}
\value{
XML version of the page containing links to
        manuals.
}
\description{
Current location of the DHS recode manuals.
}
