% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-metadata.R
\name{get_value_code}
\alias{get_value_code}
\title{Get value code matching a search string}
\usage{
get_value_code(s, data, column, ...)
}
\arguments{
\item{s}{string to search for}

\item{data}{data frame with embeded column labels}

\item{...}{additional arguments to regex search.}
}
\description{
Get value code matching a search string
}
