% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-metadata.R
\name{get_value_labels}
\alias{get_value_labels}
\title{Get value label matching a value code}
\usage{
get_value_labels(c, data, column)
}
\arguments{
\item{c}{column code to use}

\item{data}{data frame with embeded column value labels.}
}
\description{
Get value label matching a value code
}
