% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-files.R
\name{has_extension}
\alias{has_extension}
\title{Check if a string has a specific extension, fails without 
extension, returns TRUE on NULL extension.  This function
is dumb and when it encounters filenames with multiple dots
it only considers the final extension.}
\usage{
has_extension(s, e)
}
\arguments{
\item{s}{string to check for the extension}

\item{e}{string specifying the extension.}
}
\description{
Check if a string has a specific extension, fails without 
extension, returns TRUE on NULL extension.  This function
is dumb and when it encounters filenames with multiple dots
it only considers the final extension.
}
