% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-files.R
\name{index_files}
\alias{index_files}
\title{For a path return a vector of full paths to contained
file names, filtered by codes.}
\usage{
index_files(path, ...)
}
\arguments{
\item{path}{path to search for files}

\item{...}{arguments passed on to 'filter_file_names'}
}
\value{
paths to file names, limited by filtering operation
}
\description{
For a path return a vector of full paths to contained
file names, filtered by codes.
}
