% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{insert_internal_codes}
\alias{insert_internal_codes}
\title{While DHS data _usually_ encodes the survey code (country, round, and
release) in the file names, sometimes these are not correct.  It is 
possible to run a straightforward check by looking in the contained
Stata files and generating an independent label from the 'v000' column
which concatenates the country code and the DHS round (not release) 
code.  This function takes a list produced by `list_files` and 
inserts the internal country and round codes.}
\usage{
insert_internal_codes(file_index)
}
\arguments{
\item{l}{output of `list_files` function.}
}
\value{
same as input with (where possible) inserted `internal_country_code`
        and `internal_round_code` elements.
}
\description{
While DHS data _usually_ encodes the survey code (country, round, and
release) in the file names, sometimes these are not correct.  It is 
possible to run a straightforward check by looking in the contained
Stata files and generating an independent label from the 'v000' column
which concatenates the country code and the DHS round (not release) 
code.  This function takes a list produced by `list_files` and 
inserts the internal country and round codes.
}
