% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs-estimates.R
\name{jackknife_estimates}
\alias{jackknife_estimates}
\title{Produce standard DHS weighted estimates and intervals
for proportions.}
\usage{
jackknife_estimates(data, yes, count, weight, grouping = NULL)
}
\arguments{
\item{data}{data frame containing required columns.}

\item{yes}{unquoted column name containing count of 'yes' 
responses.}

\item{count}{unquoted column name containing count of 
people asked the question.}

\item{weight}{survey weight}

\item{grouping}{character vector of column names used for grouping}
}
\value{
data frame with columns for `estimate`, `sd`, `lb`, and
        `ub`, corresponding to the jackknife mean, 
        standard deviation, mean - 2 * sd, and mean + 2 * sd
}
\description{
Produce standard DHS weighted estimates and intervals
for proportions.
}
