% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs-estimates.R
\name{jackknife_weighted_mean}
\alias{jackknife_weighted_mean}
\title{Weighted mean with jackknife.}
\usage{
jackknife_weighted_mean(x, w)
}
\arguments{
\item{x}{values to average}

\item{w}{weights}
}
\value{
jackknife mean estimate
}
\description{
Weighted mean with jackknife.
}
