% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs-estimates.R
\name{jackknife_weighted_sd}
\alias{jackknife_weighted_sd}
\title{Weighted sd with jackknife.}
\usage{
jackknife_weighted_sd(x, w)
}
\arguments{
\item{x}{values to average}

\item{w}{weights}
}
\value{
jackknife sd estimate.
}
\description{
Weighted sd with jackknife.
}
