% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-files.R
\name{null_or_compare}
\alias{null_or_compare}
\title{Allow NULL to evaluate as TRUE for filtering to play
nice with missing(...)}
\usage{
null_or_compare(x, y)
}
\arguments{
\item{x}{either NULL or a vector.}
}
\value{
either NULL -> TRUE or non-NULL -> self
}
\description{
Allow NULL to evaluate as TRUE for filtering to play
nice with missing(...)
}
