\name{pdhs-package}
\alias{pdhs-package}
\alias{pdhs}
\docType{package}
\title{
\packageTitle{pdhs}
}
\description{
\packageDescription{pdhs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pdhs}
\packageIndices{pdhs}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{pdhs}

Maintainer: \packageMaintainer{pdhs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
