% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKBR52.R
\name{PKBR52}
\alias{PKBR52}
\title{Pakistan Demographic and Health Survey 2006-07 Births Recode Data}
\description{

}
\examples{
# Loading PKBR52
 \dontrun{
 load(url(
  "https://github.com/myaseen208/PDHSData/raw/master/DHS2006-07/PKBR52.RData"
  ))
  PKBR52
 }
}
\references{
\enumerate{
  \item National Institute of Population Studies (NIPS), Pakistan (\url{http://nips.org.pk/Home.htm}).
 }
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  }
}
