% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKIR52.R
\name{PKIR52}
\alias{PKIR52}
\title{Pakistan Demographic and Health Survey 2006-07 Individual Recode Data}
\description{

}
\examples{
# Loading PKIR52
 \dontrun{
 load(url(
  "https://github.com/myaseen208/PDHSData/raw/master/DHS2006-07/PKIR52.RData"
  ))
  PKIR52
 }
}
\references{
\enumerate{
  \item National Institute of Population Studies (NIPS), Pakistan (\url{http://nips.org.pk/Home.htm}).
 }
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  }
}
