% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKKR52.R
\name{PKKR52}
\alias{PKKR52}
\title{Pakistan Demographic and Health Survey 2006-07 Children's Recode Data}
\description{

}
\examples{
# Loading PKKR52
 \dontrun{
 load(url(
  "https://github.com/myaseen208/PDHSData/raw/master/DHS2006-07/PKKR52.RData"
  ))
  PKKR52
 }
}
\references{
\enumerate{
  \item National Institute of Population Studies (NIPS), Pakistan (\url{http://nips.org.pk/Home.htm}).
 }
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  }
}
