% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-filenames.R
\name{process_filename}
\alias{process_filename}
\title{Specialize for just one path element, error otherwise.}
\usage{
process_filename(path)
}
\arguments{
\item{path}{path to file to extract from.}
}
\value{
a named vector with the translated long-form
        file name components.
}
\description{
Specialize for just one path element, error otherwise.
}
