% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-filenames.R
\name{process_filenames}
\alias{process_filenames}
\title{Extract all filename components into meaningful strings
using the more specific functions.}
\usage{
process_filenames(path)
}
\arguments{
\item{path}{path to file to extract from.}
}
\value{
a named vector with the translated long-form
        file name components.
}
\description{
Extract all filename components into meaningful strings
using the more specific functions.
}
