% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go-home.R
\name{static}
\alias{static}
\title{Returns the package's static resource directory (for the 
loaded version).}
\usage{
static()
}
\description{
Returns the package's static resource directory (for the 
loaded version).
}
