% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-files.R
\name{filter_file_names}
\alias{filter_file_names}
\title{Takes a path or vector of paths and filters either using
a regex pattern or based on DHS codes.  For full details
of code-based filtering see ?filter_file_name}
\usage{
filter_file_names(path, pattern = NULL, ...)
}
\arguments{
\item{path}{path or vector of paths to files to filter
out.}

\item{either}{NULL (for code based filtering) or a regex
to be used to filter out elements of the path.}
}
\value{
paths matching the conditions or a length-zero
        character vector.
}
\description{
Takes a path or vector of paths and filters either using
a regex pattern or based on DHS codes.  For full details
of code-based filtering see ?filter_file_name
}
