% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-filenames.R
\name{process_filename}
\alias{process_filename}
\title{Extract all filename components into meaningful strings
using the more specific functions.}
\usage{
process_filename(file)
}
\arguments{
\item{file}{name to extract from.}
}
\value{
a named vector with the translated long-form
        file name components.
}
\description{
Extract all filename components into meaningful strings
using the more specific functions.
}
